requirements.txt
Django==1.7.1
Pillow==2.6.1
celery==3.1.17
kombu==3.0.24
django-braces==1.8.0
django-celery==3.1.16
django-model-utils==2.2.0
django-uuidfield==0.5.0
django-storages==1.1.8
django-configurations==0.8
django-environ==0.3.0
docutils==0.11
gunicorn==19.1.1
netaddr==0.7.12
requests==2.5.1
Versio==0.2.1
